Moved generator to specific folder.
